+++ src/base_test/this_machine_op_test.cc
@@ -6,14 +6,14 @@
 int main() {
   ENTER_TEST;
 
-  int noacc               = chef::this_machine_op::num_of_available_cpu_cores();
-  int mem_total_kb        = chef::this_machine_op::mem_total_kb();
-  int mem_free_kb         = chef::this_machine_op::mem_free_kb();
-  int mem_buffers_kb      = chef::this_machine_op::mem_buffers_kb();
-  int mem_cached_kb       = chef::this_machine_op::mem_cached_kb();
-  int mem_used_kb         = chef::this_machine_op::mem_used_kb();
-  int boot_timestamp      = chef::this_machine_op::boot_timestamp();
-  int up_duration_seconds = chef::this_machine_op::up_duration_seconds();
+  int32_t noacc               = chef::this_machine_op::num_of_available_cpu_cores();
+  int32_t mem_total_kb        = chef::this_machine_op::mem_total_kb();
+  int32_t mem_free_kb         = chef::this_machine_op::mem_free_kb();
+  int32_t mem_buffers_kb      = chef::this_machine_op::mem_buffers_kb();
+  int32_t mem_cached_kb       = chef::this_machine_op::mem_cached_kb();
+  int32_t mem_used_kb         = chef::this_machine_op::mem_used_kb();
+  int64_t boot_timestamp      = chef::this_machine_op::boot_timestamp();
+  int64_t up_duration_seconds = chef::this_machine_op::up_duration_seconds();
   uint64_t in_bytes;
   uint64_t out_bytes;
   assert(chef::this_machine_op::io_bytes("notexist", &in_bytes, &out_bytes) == -1);
@@ -25,8 +25,8 @@ int main() {
   printf("mem_buffers_kb: %d\n", mem_buffers_kb);
   printf("mem_cached_kb: %d\n", mem_cached_kb);
   printf("mem_used_kb: %d\n", mem_used_kb);
-  printf("boot_timestamp: %d\n", boot_timestamp);
-  printf("up_duration_seconds: %d\n", up_duration_seconds);
+  printf("boot_timestamp: %ld\n", boot_timestamp);
+  printf("up_duration_seconds: %ld\n", up_duration_seconds);
   printf("in_bytes: %lu\n", in_bytes);
   printf("out_bytes: %lu\n", out_bytes);
